==> 6_hybrid_strategies.Rmd <==
---
title: "RTS,S strategies"
author: "Hillary Topazian"
date: "`r format(Sys.time(), '%d %B %Y')`"
output: 
  html_document:
    theme: lumen
    highlight: tango
    code_folding: hide
    toc: true
    toc_depth: 2
    toc_float: true
---
  
```{r setup, message=FALSE, warning=FALSE}
# packages
library(tidyverse)
library(sf)
library(kableExtra)
library(malariasimulation)
library(patchwork)
library(scales)
library(LaCroixColoR)
library(data.table)
library(lubridate)

# devtools::install_github('mrc-ide/malariasimulation@dev', force=TRUE)
# devtools::install_github('johannesbjork/LaCroixColoR')
knitr::opts_knit$set(message=FALSE, warning=FALSE)
```

```{css style settings, echo = FALSE}
blockquote {
    padding: 10px 20px;
    margin: 0 0 20px;
    font-size: 12px;
    border-left: 5px solid
}
```

<br>

# Objectives

- Code up a new 'hybrid' strategy: age-based programmatic roll-out of RTS,S doses 1-3 with calendar-based seasonal boosters. Compare this hybrid strategy to standard age-based and calendar-based vaccination strategies by measuring intervention efficacy across seasonal settings.

- Explore the minimum gap between receiving the 3rd dose and the first booster dose under the hybrid strategy. How many children will go through their first high transmission season un-vaccinated? 

- Explore the advantages of adding a 5th dose. 

- Other ideas: equity in coverage gaps, malariatools (Tom & Ellie), incorporating seasonal births, cost-effectiveness of ITNs vs. RTS,S.

<br>

# Set-up

Parameters used in `malariasimulation`:

  * **Seasonality**
    + High seasonality - g0 = 0.284596, g = c(-0.317878, -0.0017527, 0.116455), h = c(-0.331361, 0.293128, -0.0617547)
    + Low seasonality - g0 = 0.285505, g = c(-0.325352, -0.0109352, 0.0779865), h = c(-0.132815, 0.104675, -0.013919)
    + seasonal intervention efficacy starts at 4 months for seasonal settings and 6 months for highly seasonal settings.
  
  * **Human population** = 100,000. Average_age = 8453.323.
  
  * **Warm-up period** = 20 years for malariasimulation to stabilize at EIR equilibrium
  
  * **Simulation length** = 15 years
  
  * **Species** = 25% arabiensis, 25% funestus, 50% gambiae. 
  
  * **Treatment** 
    + drug = AL
    + drug_efficacy = 0.95000
    + drug_rel_c = 0.05094
    + drug_prophylaxis_shape = 11.30000
    + drug_prophylaxis_scale = 10.60000
    + coverage = 0.45
  
  * **RTS,S** 
    + dose 1-3 coverage = 0.80 (of target population)
    + booster coverage = 0.80 (of vaccinated population)
    + boosted antibody mean and stdev = 5.56277, 0.35 
   
  * **SMC** <span style="color: red;">curently OFF</span>
    + drug = SP-AQ
    + drug_efficacy = 0.75
    + drug_rel_c = 0.32
    + drug_prophylaxis_shape = 3.67 (modified from default)
    + drug_prophylaxis_scale = 41.48 (modified from default)
    + coverage = 0.75
    + min-age = 3 months
    + max-age = 5 years
    + synergy drug_prophylaxis_shape = 2.955348
    + synergy drug_prophylaxis_scale = 58.99686
    + 4 monthly rounds, with timing based on seasonality curves
    
  * **Outcomes**
    + unless otherwise stated events averted are calculated relative to a baseline no-vaccination scenario
    + clinical incidence calculated among individuals 0-100 years
    + severe malaria enabled
    + deaths - calculating deaths from severe disease use the following: mortality = 0.215 * severe incidence

<br>

## Link PfPR<sub>2-10 to EIR

`malariasimulation` only accepts EIR inputs using the function `set_equilibrium()`. Here I've run a range of outputs to back link PfPR<sub>2-10</sub> values to EIR. Essentially, we set our baseline parameters and then run multiple simulations at various EIRs. We then create a best-fit line through these points, and match with PfPR<sub>2-10</sub> values. Separate matching processes must take place for each seasonality setting.

**Figure 1. ** Matching PfPR<sub>2-10</sub> and EIR. The blue line in the following plots displays results from `malariaEquilibrium`, which underestimate the EIRs to be used in the model once seasonality and treatment / ITNs are in place. The table below the plot shows the EIRs which need to be used in `malariasimulation` to obtain accurate results for each PfPR<sub>2-10</sub> of interest.

``` {r prevmatch, message=FALSE, warning=FALSE, fig.width=7, fig.height=4, fig.align="center"}
# loop over malariasimulation runs
init_EIR <- c(0.01, 0.1, seq(1,9,1), seq(10, 100, by=5)) # set EIR values

# files are created in the '2_PfPR_match_eir.R' script
# files ending in F are for individual_mosquitoes = F 
# endings in T are for individual_mosquitoes = T
EIR_prev1 <- readRDS('C:/Users/htopazia/OneDrive - Imperial College London/Github/rtss_malariasimulation/rds/EIR_prev_highF.rds') %>% mutate(season = 'high seasonality')

EIR_prev2 <- readRDS('C:/Users/htopazia/OneDrive - Imperial College London/Github/rtss_malariasimulation/rds/EIR_prev_lowF.rds') %>% mutate(season = 'low seasonality')

EIR_prev <- rbind(EIR_prev1, EIR_prev2)

# calculate EIR / prev 2-10 relationship from malariEquilibrium
eir <- seq(from = .1, to = 100, by=.5)
eq_params <- malariaEquilibrium::load_parameter_set("Jamie_parameters.rds")

prev <- vapply( # calculate prevalence between 2:10 for a bunch of EIRs
  eir,
  function(eir) {
    eq <- malariaEquilibrium::human_equilibrium(
      eir,
      ft=0,
      p=eq_params,
      age=0:100
    )
    sum(eq$states[2:10, 'pos_M']) / sum(eq$states[2:10, 'prop'])
  },
  numeric(1)
)

prevmatch <- as_tibble(cbind(eir,prev))

# plot
colors <- c("malariaEquilibrium" = "blue", "malariasimulation" = "red")

p <- 
  ggplot(data=EIR_prev) + 
  geom_point(aes(x=init_EIR, y=prev), color='black') + 
  stat_smooth(aes(x=init_EIR, y=prev, color='malariasimulation'), method = 'gam', n=1000) + 
  geom_line(data=prevmatch, aes(x=eir, y=prev, color = 'malariaEquilibrium')) + 
  labs(x='initial EIR', y=expression(paste(italic(Pf),"PR"[2-10])), color='model') + 
  facet_wrap('season', nrow=1) + 
  scale_color_manual(values = colors) + 
  theme_classic()

p

# extract points from stat_smooth
p2 <- ggplot_build(p)
p2 <- p2$data[[2]]

p2 <- p2 %>% as_tibble() %>% 
  select(x,y,PANEL) %>% 
  rename(init_EIR=x, pred=y, season=PANEL) %>%
  mutate(season=ifelse(season==1,'high seasonality','low seasonality'))

# Pre-intervention baseline PfPR2-10 starting at values 10, 25, 35, 55 
PfPR <- as_tibble_col(c(.10, .25, .35, .55), column_name="pfpr")

# match via stat_smooth predictions
match <- PfPR %>%
  fuzzyjoin::difference_left_join(p2, by=c("pfpr"="pred"), max_dist=1, distance_col="dist") %>%
  group_by(pfpr, season) %>% slice_min(dist) %>% 
  rename(eir=init_EIR) %>% 
  mutate(eir = as.numeric(signif(eir,3)), seasonality=case_when(season=='high seasonality'~'highly_seasonal', TRUE~'seasonal')) %>% 
  ungroup() %>% select(-season)

# print table
match %>% select(pfpr, eir, seasonality) %>% 
  kbl(col.names = c('PfPR', 'EIR', 'seasonality'), digits=2, escape=F, align="c") %>%
  kable_styling(bootstrap_options = c("striped", "hover", fixed_thread=T), full_width = F)

```

<br>

## Plot seasonality and interventions

**Figure 2a.** Clinical incidence output from `malariasimulation` runs over highly seasonal and seasonal settings for a variety of transmission intensities. Model runs include seasonality and treatment treatment but no RTS,S or SMC interventions. Potential intervention timings are shown via vertical dashed lines. RTS,S dose #3 (and booster dose #1) efficacy starts at month 4 for seasonal settings and month 6 for highly seasonal settings. SMC rounds are spaced so that they cover the high transmission season. 

```{r seasonalitymalariasimulation, message=FALSE, warning=FALSE, fig.width=7, fig.height=4, fig.align="center"}
# read in data from HPC runs
output <- readRDS('C:/Users/htopazia/OneDrive - Imperial College London/Github/rtss_malariasimulation/rds/rtss_hybrid_all.rds')

# set parameters
human_population <- 1000
year <- 365
month <- year/12
warmup <- 4*year
sim_length <- 1*year
  
# highly seasonal parameters
high_seas <- get_parameters(list(
  human_population = human_population,
  model_seasonality = TRUE, 
  g0 = 0.284596, 
  g = c(-0.317878, -0.0017527, 0.116455),
  h = c(-0.331361, 0.293128, -0.0617547),
  incidence_rendering_min_ages = 0,
  incidence_rendering_max_ages = 100 * year,
  clinical_incidence_rendering_min_ages = 0,
  clinical_incidence_rendering_max_ages = 100 * year,
  individual_mosquitoes = FALSE))

# seasonal parameters
low_seas <- get_parameters(list(
  human_population = human_population,
  model_seasonality = TRUE, 
  g0 = 0.285505, 
  g = c(-0.325352, -0.0109352, 0.0779865),
  h = c(-0.132815, 0.104675, -0.013919),
  incidence_rendering_min_ages = 0,
  incidence_rendering_max_ages = 100 * year,
  clinical_incidence_rendering_min_ages = 0,
  clinical_incidence_rendering_max_ages = 100 * year,
  individual_mosquitoes = FALSE))

# vaccination times
first_highseas <- c((peak_season_offset(high_seas)/month) + c(-3.5,-2.5,-1.53,-1.5,-0.5,0.5,1.5)) %>% 
  as_tibble() %>%
  mutate(color = c(rep('blue',3), rep('red',4)),
         type = c(rep('RTS,S',3), rep('SMC',4)),
         seasonality = 'highly_seasonal')

first_lowseas <- c((peak_season_offset(low_seas)/month) + c(-5.5,-4.5,-3.5,-1.5,-0.5,0.5,1.5)) %>%
  as_tibble() %>%
  mutate(color = c(rep('blue',3), rep('red',4)),
         type = c(rep('RTS,S',3), rep('SMC',4)),
         seasonality = 'seasonal')

vaxtimes <- rbind(first_highseas, first_lowseas)

# post-processing - just select one year from HPC output to plot seasonality
output2 <- output %>% filter(intervention=='none') %>%
  mutate(month=ceiling(timestep/(365/12))) %>%
  filter(month<=12) %>%
  group_by(month, eir) %>%
  summarize(inc_month=sum(n_inc_clinical_0_36500 / n_0_36500, na.rm=T)) %>%
  left_join(match, by='eir') %>%
  left_join(vaxtimes, by='seasonality')

# plot
seas.labs <- c("Highly Seasonal", "Seasonal")
names(seas.labs) <- c("highly_seasonal", "seasonal")

ggplot(output2) + 
  geom_line(aes(x=month, y=inc_month, color=as.factor(pfpr))) + 
  geom_vline(data=vaxtimes[vaxtimes$type=='RTS,S',], aes(xintercept=value, fill=type), color='white', lty=2, show.legend=T) + # dummy for legend
  geom_vline(data=vaxtimes[vaxtimes$type=='RTS,S',], aes(xintercept=value, fill=type), color='blue', lty=2, show.legend=F) +
  geom_vline(data=vaxtimes[vaxtimes$type=='SMC',], aes(xintercept=value, fill=type), color='red', lty=2, show.legend=F) +
  geom_rect(data=vaxtimes[vaxtimes$type=='SMC',], aes(xmin = value, xmax = value + 1, ymin = 0, ymax = 0.2), fill="red", alpha = 0.1, show.legend = F) +
  labs(x="Timesteps (month)",
       y="Monthly clinical incidence (all ages)",
       color=expression(paste(italic(Pf),"PR"[2-10])),
       fill='intervention') + 
  scale_y_continuous(limits=c(0,0.25),breaks=c(0,0.1,0.2)) +
  scale_x_continuous(limits=c(1,12),breaks=seq(1,12,1)) + 
  scale_fill_manual('Intervention', values = c('blue', 'red'), 
       guide=guide_legend(override.aes = list(colour=c("blue", "red")), order=2),
       labels=c("RTS,S", "SMC")) + 
  facet_grid(seasonality  ~ ., labeller = labeller(seasonality = seas.labs)) + 
  theme_classic()

```

**Figure 2b.** Clinical incidence output from `malariasimulation` runs over seasonal settings for a variety of transmission intensities. Seasonality profiles match those at study study sites in [Burkina Faso and Mali](https://journals.plos.org/plosmedicine/article?id=10.1371/journal.pmed.1003214). Model runs include seasonality and treatment but no RTS,S or SMC interventions. 

The map below shows Matt C's study sites. I took rainfall data (Jan 2015 to Dec 2016) from the admin units where the study sites were located. The admin1 units are colored in blue and the admin2 units are outlined in black. Rainfall patterns were similar when run with admin1 and admin2. 

```{r mali BF, message=FALSE, warning=FALSE, fig.width=7, fig.height=4, fig.align="center"}
smcdat <- read_csv('./LSHTM_sites_SMC.csv') %>% 
  filter(seasonality_no_smc == 1 & country != 'BOTH') %>%
  group_by(country) %>%
  mutate(model=case_when(country=='BF'~"burkina faso", 
                         country=='ML'~'mali'),
         inc_month=rate_1000_personmonths / 1000,
         date=mdy(myr),
         year=year(date),
         month=month(date)) 

smc15 <- smcdat %>% filter(year==2015)
smc16 <- smcdat %>% filter(year==2016)

# ggplot() + 
#   geom_line(data = smc15, aes(x=month, y=inc_month, color=model)) + 
#   geom_line(data = smc16, aes(x=month, y=inc_month, color=model))

# set parameters
human_population <- 10000
year <- 365
month <- year/12
warmup <- 4*year
sim_length <- 1*year

# values from 2015-2016 
# admin 1
# Hauts-Bassins (BF): 3.050911 -3.635848 -0.08368039 1.0269266 -2.564021 1.956763 -0.1073778
# Sikasso (M): 2.961674 -3.587095 0.18946440 0.6632403 -2.423344 1.764444 -0.2240691
# admin 2
# Hauts-Bassins (BF): 2.910332 -3.497756 -0.07166861 1.0143044 -2.476774 1.953536 -0.1429758
# Sikasso (M): 2.960722 -3.615782 0.30245314 0.4661306 -2.283491 1.546425 -0.1387256

# highly seasonal parameters
burkina_faso <- get_parameters(list(
  human_population = human_population,
  model_seasonality = TRUE, 
  g0 = 2.910332, 
  g = c(-3.497756, -0.07166861, 1.0143044),
  h = c(-2.476774, 1.953536, -0.1429758),
  incidence_rendering_min_ages = 0,
  incidence_rendering_max_ages = 100 * year,
  clinical_incidence_rendering_min_ages = c(0,0),
  clinical_incidence_rendering_max_ages = c(5 * year, 100*year),
  individual_mosquitoes = TRUE))

# seasonal parameters
mali <- get_parameters(list(
  human_population = human_population,
  model_seasonality = TRUE, 
  g0 = 2.960722, 
  g = c(-3.615782, 0.30245314, 0.4661306),
  h = c(-2.283491, 1.546425, -0.1387256),
  clinical_incidence_rendering_min_ages = c(0,0),
  clinical_incidence_rendering_max_ages = c(5 * year, 100*year),
  individual_mosquitoes = TRUE))

year <- 365
month <- year/12

runsim <- function(params, starting_EIR, warmup, sim_length, season){
  year <- 365
  month <- year/12
  
  params <- set_species(params, species = list(arab_params, fun_params, gamb_params), 
                        proportions = c(0.25, 0.25, 0.5))
  params$phi_bednets <- c(0.9, 0.9, 0.89) 
  params$phi_indoors <- c(0.96, 0.98, 0.97)
  
  params <- set_drugs(params, list(AL_params))
  params <- set_clinical_treatment(params, 1, c(1), c(0.45))
  params <- set_equilibrium(params, starting_EIR)
  
  run_simulation(warmup + sim_length, params) %>% 
    mutate(eir=starting_EIR, 
           timestep=timestep-warmup,
           model=season) %>%
    filter(timestep > 0)
}

output1 <- runsim(mali, 5, warmup, sim_length, 'mali')
output2 <- runsim(mali, 10, warmup, sim_length, 'mali')
output3 <- runsim(mali, 15, warmup, sim_length, 'mali')
output4 <- runsim(mali, 20, warmup, sim_length, 'mali')
output5 <- runsim(burkina_faso, 5, warmup, sim_length, 'burkina faso')
output6 <- runsim(burkina_faso, 10, warmup, sim_length, 'burkina faso')
output7 <- runsim(burkina_faso, 15, warmup, sim_length, 'burkina faso')
output8 <- runsim(burkina_faso, 20, warmup, sim_length, 'burkina faso')
  
dat <- rbindlist(list(output1, output2, output3, output4, output5, output6, output7, output8), fill = TRUE, idcol="file", use.names = T)

# vaccination times
first_bf <- c((peak_season_offset(burkina_faso)/month) + c(-3,-2,-1,-1.5,-0.5,0.5,1.5)) %>% 
  as_tibble() %>%
  mutate(color = c(rep('blue',3), rep('red',4)),
         type = c(rep('RTS,S',3), rep('SMC',4)),
         seasonality = 'burkina faso')

first_m <- c((peak_season_offset(mali)/month) + c(-3,-2,-1,-1.5,-0.5,0.5,1.5)) %>%
  as_tibble() %>%
  mutate(color = c(rep('blue',3), rep('red',4)),
         type = c(rep('RTS,S',3), rep('SMC',4)),
         seasonality = 'mali')

vaxtimes2 <- rbind(first_bf, first_m)

# post-processing - just select one year from HPC output to plot seasonality
dat2 <- dat %>% 
  mutate(month=ceiling(timestep/(365/12))) %>%
  filter(month<=12) %>%
  group_by(month, eir, model) %>%
  summarize(inc_month=sum(n_inc_clinical_0_1825 / n_0_1825, na.rm=T)) %>%
  left_join(vaxtimes2, by=c('model'='seasonality'))

# plot
ggplot(dat2) + 
  geom_line(aes(x=month, y=inc_month, color=as.factor(eir))) + 
  # geom_vline(data=. %>% filter(type=='RTS,S'), aes(xintercept=value, fill=type), color='white', lty=2, show.legend=T) + # dummy for legend
  # geom_vline(data=. %>% filter(type=='RTS,S'), aes(xintercept=value, fill=type), color='blue', lty=2, show.legend=F) +
  # geom_vline(data=. %>% filter(type=='SMC'), aes(xintercept=value, fill=type), color='red', lty=2, show.legend=F) +
  # geom_rect(data=. %>% filter(type=='SMC'), aes(xmin = value, xmax = value + 1, ymin = 0, ymax = 0.5), fill="red", alpha = 0.005, show.legend = F) +
  geom_line(data=smc15, aes(x=month, y=inc_month), color='black') + 
  geom_line(data=smc16, aes(x=month, y=inc_month), color='black') + 
  labs(x="Timesteps (month)",
       y="Monthly clinical incidence (0-5 years)",
       color='EIR',
       fill='intervention') + 
  scale_y_continuous(limits=c(0,0.5),breaks=c(0,0.1,0.2,0.3,0.4)) +
  scale_x_continuous(limits=c(1,12),breaks=seq(1,12,1)) + 
  scale_fill_manual('Intervention', values = c('blue', 'red'), 
       guide=guide_legend(override.aes = list(colour=c("blue", "red")), order=2),
       labels=c("RTS,S", "SMC")) + 
  facet_grid(model ~ .) + 
  theme_classic()

# study sites
admin0 <- readRDS("M:/Eradication/rds/admin0.RDS") %>% filter(Country %in% c('Mali', 'Burkina Faso', 'Senegal', 'Guinea', 'Mauritania', 'Algeria', 'Niger', "CÃ´te d'Ivoire", 'Ghana', 'Togo', 'Benin', 'Nigeria', 'Sierra Leone', 'Morocco'))
admin1 <- readRDS("M:/Eradication/rds/admin1.RDS") %>% filter(admin1 %in% c('Hauts-Bassins', 'Sikasso'))
countries <- filter(admin0, Country %in% c('Mali', 'Burkina Faso')) 
admin2 <- readRDS("M:/Eradication/rds/admin2.RDS") %>% 
  filter(admin2 %in% c('Tuy', 'Bougouni'))

# plot
ggplot() + 
  geom_sf(data = admin0, fill="cornsilk2",color="cornsilk3") +
  geom_sf(data = countries, fill="cornsilk", color="tan4", size=0.5) +
  geom_sf(data = admin1, aes(fill = Country)) + 
  geom_sf(data = admin2, color='black', fill=NA) + 
  labs(fill = 'country sites', x = "", y = "") + 
  theme_bw() + 
  scale_x_continuous(limits=c(-12, 4)) + 
  scale_y_continuous(limits=c(9.5, 25)) + 
  theme(axis.text.x = element_blank(), axis.text.y = element_blank(), axis.ticks = element_blank())


```

<br> 

# 1. Severe cases

**Figure 3.** Distribution of severe case counts by age and transmission setting. Age represents the upper end of a 1 year age-group. E.g. age "5" represents children ages 4-5. Results are drawn from a baseline model which includes no RTS,S or SMC intervention.

```{r age dist severe, message=FALSE, warning=FALSE, fig.width=7, fig.height=3, fig.align="center"}
# read in data from HPC runs
output <- readRDS('C:/Users/htopazia/OneDrive - Imperial College London/Github/rtss_malariasimulation/rds/rtss_severe_all.rds') %>% 
  group_by(model, eir, age_low, age_high) %>%
  summarize(across(s_cases:s_inc, mean, na.rm=T)) %>%
  left_join(match)


# plot
ggplot(output) + 
  geom_line(aes(x=age_high, y=s_cases, color=as.factor(pfpr)), size=0.9) + 
  facet_wrap(.~seasonality, labeller = labeller(seasonality = seas.labs)) + 
  labs(x = "Age (years)",
       y = "Severe cases",
       color = expression(paste(italic(Pf),"PR"[2-10])),
       parse=TRUE) + 
  scale_x_continuous(breaks = seq(0,20,1), labels = c(0, rep("",4), "5", rep("",4), "10", rep("",4), "15", rep("",4), "20") ) + 
  theme_classic() + 
  coord_cartesian(xlim=c(0,20))

```

<br>

**Table 1.** Distribution of severe case counts by age and transmission setting. Results are drawn from a baseline model which includes no RTS,S or SMC intervention. Blue cells highlight the age group with the maximum number of severe cases in each seasonality and PfPR<sub>2-10</sub> setting. 

```{r age dist severe 2, message=FALSE, warning=FALSE, fig.width=7, fig.height=3, fig.align="center"}
# summary table
table <- output %>% ungroup() %>% 
  filter(age_high <= 10) %>%
  group_by(model, pfpr) %>% 
  mutate(age = paste0(age_low, '-', age_high), 
         s_cases = round(s_cases), 
         s_cases = cell_spec(s_cases, background=ifelse(s_cases == max(s_cases), "lightblue", 'white')),
         model = ifelse(model=='high_seas', "highly seasonal", "seasonal")) %>%
  dplyr::select(model, pfpr, age, s_cases) %>%
  pivot_wider(names_from='age', values_from='s_cases') 

x <- table %>% 
  kbl(col.names = c('seasonality','PfPR','0-1','1-2','2-3','3-4','4-5','5-6','6-7','7-8','8-9','9-10'), escape=F, align="c") %>%
  kable_styling(bootstrap_options = c("striped", "hover", fixed_thread=T), full_width = F)

add_header_above(x, c("", "", "severe case counts by age group"=10))

```

<br>

**Figure 4.** Distribution of severe case counts by age and transmission setting. Potential ntervention timings are shown via vertical dashed lines. RTS,S dose #3 (and booster dose #1) efficacy starts at month 4 for seasonal settings and month 6 for highly seasonal settings. SMC rounds are spaced so that they cover the high transmission season. Results are drawn from a baseline model which includes no RTS,S or SMC intervention.

```{r seasonal severe, message=FALSE, warning=FALSE, fig.width=7, fig.height=3, fig.align="center"}
# read in data from HPC runs
output <- readRDS('C:/Users/htopazia/OneDrive - Imperial College London/Github/rtss_malariasimulation/rds/rtss_hybrid_all.rds')

# data manipulation
output2 <- output %>%
  filter(intervention == 'none') %>%
  mutate(month=ceiling(timestep/(365/12))) %>%
  filter(month<=12) %>%
  group_by(model, month, eir) %>%
  # summarize cases in under 5s
  summarize(severe=sum(n_inc_severe_0_1825, na.rm=T)) %>% ungroup() %>%
  left_join(match) %>%
  left_join(vaxtimes, by='seasonality')

ggplot(output2) + 
  geom_line(aes(x=month, y=severe, color=as.factor(pfpr))) + 
  geom_vline(data=vaxtimes[vaxtimes$type=='RTS,S',], aes(xintercept=value, fill=type), color='white', lty=2, show.legend=T) + # dummy for legend
  geom_vline(data=vaxtimes[vaxtimes$type=='RTS,S',], aes(xintercept=value, fill=type), color='blue', lty=2, show.legend=F) +
  geom_vline(data=vaxtimes[vaxtimes$type=='SMC',], aes(xintercept=value, fill=type), color='red', lty=2, show.legend=F) +
  geom_rect(data=vaxtimes[vaxtimes$type=='SMC',], aes(xmin = value, xmax = value + 1, ymin = 0, ymax = 180), fill="red", alpha = 0.1, show.legend = F) +
  labs(x="Timesteps (month)",
       y="Monthly severe cases (0-5 years)",
       color=expression(paste(italic(Pf),"PR"[2-10])),
       fill='intervention') + 
  scale_y_continuous() +
  scale_x_continuous(limits=c(1,12),breaks=seq(1,12,1)) + 
  scale_fill_manual('Intervention', values = c('blue', 'red'), 
       guide=guide_legend(override.aes = list(colour=c("blue", "red")), order=2),
       labels=c("RTS,S", "SMC")) + 
  facet_grid(seasonality  ~ ., labeller = labeller(seasonality = seas.labs)) + 
  theme_classic()

```

<br> 

# 2. RTS,S strategies

**Table 2.** Description of RTS,S implementation strategies.

```{r intervention table, message=FALSE, warning=FALSE}
# table of RTS,S strategies - manual input
scenario <- c("Age-based vaccination", "Calendar-based vaccination", "Hybrid vaccination")

dose13 <- c("6, 7.5, and 9 months of age", "1, 2, and 3 months before the high transmission season, among those 5-17 months of age", "6, 7.5, and 9 months of age")

dose4 <- c("27 months of age", "12 months following dose 3", "One month before the high transmission season")
  
dose5 <- c("-", "12 months following dose 4 (optional)", "12 months following dose 4 (optional)")
  
notes <- c("-", "-", "Vary the minimum wait time between dose 3 and dose 4")

# create table with kable()
x <- knitr::kable(cbind(scenario, dose13, dose4, dose5, notes), "html", align=rep('l', 5), col.names = c("scenario", "doses 1-3", "dose 4", "dose 5", "additional analyses")) %>% row_spec(0, align = "c") %>% kable_styling(bootstrap_options = c("striped", "hover", fixed_thread=T), full_width = F)

add_header_above(x, c(" ", "Vaccination schedule" = 3, " "))

```

<br>

# 2. Cases

**Figure 5a.** The number of clinical cases per year among children 0-5 years under each intervention strategy, stratified by season and PfPR<sub>2-10</sub> transmission setting. 

```{r compare, message=FALSE, warning=FALSE, fig.width=8, fig.height=6, fig.align="center"}
# read in data from HPC runs
output <- readRDS('C:/Users/htopazia/OneDrive - Imperial College London/Github/rtss_malariasimulation/rds/rtss_hybrid_all.rds')

# data manipulation
output2 <- output %>%
  filter(intervention != 'hybrid' | minwait == 30) %>%
  mutate(year=ceiling(timestep/365)) %>%
  group_by(intervention, model, year, eir, fifth) %>%
  # summarize cases in under 5s
  summarize(cases = sum(p_inc_clinical_0_1825, na.rm=T),
            severe = sum(p_inc_severe_0_1825, na.rm=T),
            deaths = sum(0.215 * p_inc_severe_0_1825, na.rm=T)) %>% ungroup() %>%
  left_join(match) %>%
  # attach PfPRs
  mutate(pfpr = factor(pfpr, levels = c(0.1, 0.25, 0.35, 0.55), labels = c('PfPR 10', 'PfPR 25', 'PfPR 35', 'PfPR 55' ))) %>%
  # label interventions
  mutate(intervention = case_when(intervention == 'EPI' ~ 'age-based',
             intervention == 'none' ~ 'none',
             intervention == 'SV 4-dose' ~ 'calendar-based 4-dose',
             intervention == 'SV 5-dose' ~ 'calendar-based 5-dose',
             intervention == 'hybrid' & fifth == 0 ~ 'hybrid 4-dose',
             intervention == 'hybrid' & fifth == 1 ~ 'hybrid 5-dose'),
         intervention = factor(intervention, levels = c('none', 'age-based', 'calendar-based 4-dose', 'calendar-based 5-dose', 'hybrid 4-dose', 'hybrid 5-dose')))

# plot
ggplot(data=output2, aes(x=year, y=cases, group=interaction(pfpr, intervention), color=intervention)) + 
  geom_line() + 
  scale_y_continuous(labels = comma) + 
  scale_x_continuous(breaks = seq(1,15,1), limits = c(0,20)) + 
  facet_grid(. ~ seasonality, labeller = labeller(seasonality = seas.labs)) + 
  labs(x='Year', y='Clinical cases (0-5 years)', color='Intervention') +
  theme_classic() + 
  theme(axis.text.x = element_text(angle = 90, hjust=1, vjust=0.5)) + 
  annotate("text", x = 18, y = 4000, label = 'PfPR 10') + 
  annotate("text", x = 18, y = 13000, label = 'PfPR 25') + 
  annotate("text", x = 18, y = 22000, label = 'PfPR 35') + 
  annotate("text", x = 18, y = 41000, label = 'PfPR 55') 

```

<br>

**Figure 5b.** The number of clinical cases per year among children 0-5 years under each intervention strategy, stratified by season and PfPR<sub>2-10</sub> transmission setting. 

```{r compare2, message=FALSE, warning=FALSE, fig.width=8, fig.height=6, fig.align="center"}

ggplot(data=output2, aes(x=year, y=cases, group=interaction(pfpr, intervention), color=intervention)) + 
  geom_line() + 
  scale_y_continuous(labels = comma) + 
  scale_x_continuous(breaks = seq(1,15,1), limits = c(0,20)) + 
  facet_grid(pfpr ~ seasonality, labeller = labeller(seasonality = seas.labs), scales='free') + 
  labs(x='Year', y='Clinical cases (0-5 years)', color='Intervention') +
  theme_classic() + 
  theme(axis.text.x = element_text(angle = 90, hjust=1, vjust=0.5)) 

```

<br>

**Figure 6.** The number of clinical cases, severe cases, and deaths averted among children 0-5 years over the whole 15 simulation period, stratified by intervention. The reference group is a block of scenarios run with no RTS,S intervention. 

```{r compare3, message=FALSE, warning=FALSE, fig.width=8, fig.height=10, fig.align="center"}
# data manipulation
output3 <- output2 %>% 
  group_by(intervention,seasonality, pfpr) %>%  
  summarize(cases = sum(cases, na.rm=T),
            severe = sum(severe, na.rm=T),
            deaths = sum(deaths, na.rm=T)) %>% ungroup()

none <- output3 %>% filter(intervention=='none') %>% 
  rename(cases0 = cases, severe0 = severe, deaths0 = deaths) %>%
  select(-intervention)

output4 <- output3 %>% filter(intervention!='none') %>% 
  left_join(none, by=c('seasonality','pfpr')) %>% 
  mutate(casesA=cases0-cases, severeA=severe0-severe, deathsA=deaths0-deaths)

# plot cases, severe cases, and deaths averted
compareplot <- function(data, var, title){ggplot(data=data, aes(x=factor(pfpr), y=var, group=intervention, fill=intervention)) + 
  geom_col(position = 'dodge') + 
  facet_grid(seasonality  ~ ., labeller = labeller(seasonality = seas.labs)) + 
  labs(x=expression(paste(italic(Pf),"PR"[2-10])), y=title, title=title, fill='Intervention') +
  theme_classic() 
}

compareplot(output4, output4$casesA, 'Clinical cases (children 0-5 years)') / 
  compareplot(output4, output4$severeA, 'Severe cases (children 0-5 years)') / 
  compareplot(output4, output4$deathsA, 'Deaths (children 0-5 years)') + plot_layout(guides = "collect") 

```


<br>

# 3. Minimum wait time between 3rd and 4th dose

**Figure 7.** Exploring the minimum wait time between receiving the 3rd dose and the first booster dose under the hybrid strategy. Until now we have set this minimum gap at 1 month. The outcome is the number of clinical cases among children 0-5 years, stratified by PfPR<sub>2-10</sub> and seasonality. The transparency of each line represents the minimum wait time, ranging from 0 to 11 months.

```{r wait time, message=FALSE, warning=FALSE, fig.width=8, fig.height=5, fig.align="center"}
# read in data from HPC runs
output <- readRDS('C:/Users/htopazia/OneDrive - Imperial College London/Github/rtss_malariasimulation/rds/rtss_hybrid_all.rds') 

# plot: minimum wait times by prfpr
output2 <- output %>% filter(grepl('wait', file)) %>% 
  filter(fifth==0) %>%                                    # limiting it to 4 doses for now
  mutate(year=ceiling(timestep/365), minwait=round(minwait/month)) %>%
  group_by(minwait, eir, model, year) %>%
  summarize(cases = sum(p_inc_clinical_0_1825, na.rm=T)) %>% ungroup() %>%
  left_join(match)

seas.labs <- c("Highly Seasonal", "Seasonal")
names(seas.labs) <- c("highly_seasonal", "seasonal")

ggplot(data=output2, aes(x=year, y=cases, group=interaction(eir,minwait), color=factor(pfpr), alpha=factor(minwait))) + 
  geom_line() + 
  labs(x = expression(paste(italic(Pf),"PR"[2-10])), 
       y = " ",
       parse=TRUE) + 
  scale_y_continuous(labels = comma) + 
  scale_x_continuous(breaks = seq(1,15,1)) + 
  facet_wrap(seasonality  ~ ., labeller = labeller(seasonality = seas.labs)) +
  labs(x='Year', y='Clinical cases among children 0-5 years', 
       color=expression(italic(PfPR[2-10]~(symbol("\045")))), alpha='Minimum wait time') +
  theme_classic()
```

<br> 

**Table 3.** Difference in the number of clinical cases as the minimum wait time varies between receiving the 3rd dose and the first booster dose under the hybrid strategy. The outcome is the number of clinical cases among children 0-5 years over the 15-year simulation period, stratified by PfPR<sub>2-10</sub> and seasonality. The reference group is a minimum wait time of 0 months. Blue cells highlight the minimum wait time resulting in the most cases averted.  

```{r wait time2, message=FALSE, warning=FALSE, fig.width=8, fig.height=6, fig.align="center"}
# table from plot ^ 
x <- output2 %>% 
  group_by(minwait, seasonality, pfpr) %>%
  summarize(cases = round(sum(cases, na.rm=T))) %>%
  arrange(seasonality, pfpr, minwait) %>%
  group_by(seasonality, pfpr) %>%
  mutate(diff = round((cases - first(cases))),
         cases = ifelse(diff==0, cases, NA),
         cases = mean(cases, na.rm=T),
         diff = cell_spec(diff, background=ifelse(diff == min(diff), "lightblue", 'white')),
         diff = ifelse(diff==0,'-',diff),
         seasonality = ifelse(seasonality=='highly_seasonal', "highly seasonal", "seasonal")) %>%
  pivot_wider(names_from = 'minwait', values_from = 'diff') %>%
  kbl(col.names = c('seasonality', 'PfPR', 'baseline clinical cases', '0','1','2','3','4','5','6','7','8','9','10','11'), escape = F, align = "c") %>%
  kable_styling(bootstrap_options = c("striped", "hover", fixed_thread = T)) 

add_header_above(x, c(rep('',3), 'minimum wait time (months)'=12)) 

```

<br>

# 4. Fifth dose

**Figure 8.** The number of doses given out under each strategy over the first 36 months. Data are limited to PfPR<sub>2-10</sub> 0.55 settings, and the hybrid strategy shown is run with a minimum wait time of 1 month between the 3rd dose of the primary series and the 1st booster dose. Seasonality curves are plotted in the background and represent curves with no RTS,S intervention.

```{r fifth dose, message=FALSE, warning=FALSE, fig.width=10, fig.height=6, fig.align="center"}
# read in data from HPC runs
output <- readRDS('C:/Users/htopazia/OneDrive - Imperial College London/Github/rtss_malariasimulation/rds/rtss_hybrid_all.rds')

# data manipulation
output2 <- output %>% 
  mutate(month=ceiling(timestep/(365/12)), minwait=round(minwait/(365/12))) %>%
  filter(is.na(minwait) | minwait == 1) %>%
  group_by(intervention, model, eir, minwait, month, fifth) %>%
  summarize(dose1 = sum(dose1, na.rm=T),
            dose2 = sum(dose2, na.rm=T),
            dose3 = sum(dose3, na.rm=T),
            dose4 = sum(dose4, na.rm=T),
            dose5 = sum(dose5, na.rm=T)) %>%
  arrange(model, eir, minwait, month) %>%
  pivot_longer(cols=c(dose1:dose5), names_to = 'dose', values_to = 'nvax') %>%
  group_by(intervention, model, eir, minwait, dose, fifth) %>%
  mutate(nvaxcum = cumsum(nvax)) %>%
  ungroup() %>%
  left_join(match) %>%
  mutate(intervention = case_when(intervention == 'EPI' ~ 'age-based',
         intervention == 'none' ~ 'none',
         intervention == 'SV 4-dose' ~ 'calendar-based 4-dose',
         intervention == 'SV 5-dose' ~ 'calendar-based 5-dose',
         intervention == 'hybrid' & fifth == 0 ~ 'hybrid 4-dose',
         intervention == 'hybrid' & fifth == 1 ~ 'hybrid 5-dose'),
         intervention = factor(intervention, levels = c('none', 'age-based', 'calendar-based 4-dose', 'calendar-based 5-dose', 'hybrid 4-dose', 'hybrid 5-dose'))) 

# total vaccinations
total <- output2 %>% ungroup() %>% 
  group_by(intervention, seasonality, eir, pfpr, month, minwait, fifth) %>%
  summarize(nvax = sum(nvax, na.rm=T)) %>%
  group_by(intervention, seasonality, eir) %>%
  mutate(nvaxcum = cumsum(nvax))

# seasonality curve
scurve <- output %>% filter(intervention=='none') %>% ungroup() %>%
  mutate(month=ceiling(timestep/(365/12))) %>%
  filter(month<=36) %>%
  group_by(month, eir) %>%
  summarize(inc_month=sum(n_inc_clinical_0_36500 / n_0_36500, na.rm=T)) %>%
  left_join(match, by='eir') %>% filter(pfpr==0.55) 

# to see default ggplot colors:
# library(scales)
# show_col(hue_pal()(5))

cols <- c('#F8766D','#A3A500','#00BF7D','#00B0F6','#E76BF3',"#000000")

# plot
ggplot(data=output2 %>% filter(month<=36 & pfpr==0.55) %>% 
  filter(intervention %in% c('none','age-based','calendar-based 5-dose') | (minwait==1 & fifth==1))) + 
  geom_line(data=scurve, aes(x=month, y=inc_month*200000), color='grey') + 
  geom_area(data=scurve, aes(x=month, y=inc_month*200000), fill='grey', alpha=0.2, outline.type='upper') + 
   geom_line(data=total %>% filter(month<=36 & pfpr==0.55) %>% 
  filter(intervention %in% c('none','age-based','calendar-based 5-dose') | (minwait==1 & fifth==1)), aes(x=month, y=nvaxcum, color='total'), size=1) + 
   geom_line(aes(x=month, y=nvaxcum, group=interaction(dose, eir), color=dose), size=1) + 
  scale_y_continuous(labels = comma, sec.axis = sec_axis(~ ./200000, name = "Monthly clinical incidence")) + # 
  scale_x_continuous(breaks=seq(0,36,6)) + 
  facet_grid(seasonality ~ intervention, labeller = labeller(seasonality = seas.labs)) +
  labs(x='Month', y='Number of vaccinations', color='# doses') +
   theme_classic() + 
   theme(plot.margin=unit(c(1,5,1,1), "pt"), axis.title.y.right = element_text(vjust=3)) +
   scale_color_manual(values = cols)

```

<br>

**Figure 9.** Comparison of the calendar-based and hybrid strategies under a 4-dose vs. 5-dose regimen. The outcome plotted is the number of clinical cases averted (as compared to no intervention) among children 0-5 years under each strategy per year. 

```{r fifth dose2, message=FALSE, warning=FALSE, fig.width=8, fig.height=6, fig.align="center"}
# calculate baseline cases
none <- output %>% filter(grepl('none', file)) %>% 
  mutate(year=ceiling(timestep/365)) %>%
  group_by(eir, model, year) %>%
  summarize(cases_base = sum(p_inc_clinical_0_1825, na.rm=T)) %>% ungroup() 
  
# data manipulation
output3 <- output %>% 
  # limit data set to SV and hybrid scenarios and rename
  filter(grepl('SV', intervention) | (grepl('hybrid', intervention) & minwait == 30)) %>%
  mutate(intervention = case_when(intervention == 'SV 4-dose' ~ 'calendar-based 4-dose',
            intervention == 'SV 5-dose' ~ 'calendar-based 5-dose',
            intervention == 'hybrid' & fifth == 0 ~ 'hybrid 4-dose',
            intervention == 'hybrid' & fifth == 1 ~ 'hybrid 5-dose')) %>%
  # summarize cases by year
  mutate(year=ceiling(timestep/365)) %>%
  group_by(intervention, eir, model, year) %>%
  summarize(cases = sum(p_inc_clinical_0_1825, na.rm=T)) %>% ungroup() %>%
  # add in pfpr and outcomes under no RTS,S
  left_join(match) %>%
  left_join(none) %>%
  # calculate cases / deaths averted
  mutate(case_avert = cases_base - cases)

# plot
ggplot(data=output3, aes(x=year, y=case_avert, group=interaction(pfpr, intervention), color=intervention)) + 
  geom_line() + 
  scale_y_continuous(labels = comma) + 
  scale_x_continuous(breaks = seq(1,15,1), limits = c(0,20)) + 
  facet_grid(. ~ seasonality, labeller = labeller(seasonality = seas.labs)) + 
  labs(x='Year', y='Clinical cases averted among children 0-5 years', color='Intervention') +
  theme_classic() + 
  theme(axis.text.x = element_text(angle = 90, hjust=1, vjust=0.5)) + 
  annotate("text", x = 18, y = 1000, label = 'PfPR 10') +
  annotate("text", x = 18, y = 3500, label = 'PfPR 25') +
  annotate("text", x = 18, y = 5000, label = 'PfPR 35') +
  annotate("text", x = 18, y = 7000, label = 'PfPR 55')
```

<br>

**Table 4.** Comparison of the calendar-based and hybrid strategies under a 4-dose vs. 5-dose regimen. The outcome is the number of clinical cases averted (as compared to no intervention) and the number of clinical cases averted per dose among children 0-5 years under each strategy over all 15 years of the simulation. Blue cells highlight the maximum cases averted or the maximum cases averted per dose in each seasonality and PfPR<sub>2-10</sub> setting. 

```{r fifth dose3, message=FALSE, warning=FALSE, fig.width=8, fig.height=4, fig.align="center"}
# total vaccinations
total2 <- total %>% ungroup() %>% 
  group_by(intervention, seasonality, pfpr) %>%
  summarize(nvax = sum(nvax, na.rm=T)) 

# data manipulation
output4 <- output3 %>% 
  group_by(intervention, seasonality, pfpr) %>%  
  summarize(case_avert = sum(case_avert, na.rm=T)) %>% 
  left_join(total2) %>%
  group_by(seasonality, pfpr) %>%
  mutate(dose = case_when(grepl('4',intervention)~4,
                          grepl('5',intervention)~5),
         intervention = case_when(grepl('hybrid',intervention)~'hybrid',
                                  grepl('calendar',intervention)~'calendar-based'),
         case_avert = round(case_avert),
         case_avert_vax = round(case_avert / nvax, 2),
         case_avert = cell_spec(case_avert, background=ifelse(case_avert == max(case_avert), "lightblue", 'white')),
         case_avert_vax = cell_spec(case_avert_vax, background=ifelse(case_avert_vax == max(case_avert_vax), "lightblue", 'white')),
         seasonality=ifelse(seasonality=='highly_seasonal', 'highly seasonal', 'seasonal')) %>%
  select(-nvax) %>%
  pivot_wider(names_from=c(intervention, dose), values_from=c(case_avert, case_avert_vax)) 

# print table
x <- output4 %>%
  kbl(col.names = c('seasonality', 'PfPR', 'calendar-based dose 4', 'calendar-based dose 5', 'hybrid dose 4', 'hybrid dose 5', 'calendar-based dose 4', 'calendar-based dose 5', 'hybrid dose 4', 'hybrid dose 5'), escape = F, align = "c") %>%
  kable_styling(bootstrap_options = c("striped", "hover", fixed_thread = T)) 

add_header_above(x, c(rep('',2), 'cases averted'=4, 'cases averted per dose'=4)) 

```

<br>

# 5. Coverage (% vaccinated)

Change figure - we set the coverage right now so this is out of children 0-5 years, but not very helpful. 

Check with Giovanni: 

- Can we set up a way to output the number of kids tagged as vaccinated within an age group for each time step? 

- See if we can calculate vaccine indirect effects


```{r coverage, message=FALSE, warning=FALSE, fig.width=8, fig.height=4, fig.align="center"}
# read in data from HPC runs
output <- readRDS('C:/Users/htopazia/OneDrive - Imperial College London/Github/rtss_malariasimulation/rds/rtss_hybrid_all.rds') 

# data manipulation
output2 <- output %>%
  filter(intervention != 'hybrid' & intervention != 'none' | minwait == 30) %>%
  mutate(year=ceiling(timestep/365)) %>%
  left_join(match) %>%
  mutate(intervention = case_when(intervention == 'EPI' ~ 'EPI',
             intervention == 'SV 4-dose' ~ 'seasonal 4-dose',
             intervention == 'SV 5-dose' ~ 'seasonal 5-dose',
             intervention == 'hybrid' & fifth == 0 ~ 'hybrid 4-dose',
             intervention == 'hybrid' & fifth == 1 ~ 'hybrid 5-dose'),
         intervention = factor(intervention, levels = c('EPI', 'seasonal 4-dose', 'seasonal 5-dose', 'hybrid 4-dose', 'hybrid 5-dose'))) %>%
  group_by(intervention, seasonality, pfpr, year) %>%
  summarize(dose1 = sum(dose1, na.rm=T) / mean(n_0_1825),
         dose2 = sum(dose2, na.rm=T) / mean(n_0_1825),
         dose3 = sum(dose3, na.rm=T) / mean(n_0_1825),
         dose4 = sum(dose4, na.rm=T) / mean(n_0_1825),
         dose5 = sum(dose5, na.rm=T) / mean(n_0_1825)) %>%
  pivot_longer(cols=c(dose1:dose5), names_to = 'dose', values_to = 'coverage') %>%
  arrange(intervention, seasonality, pfpr, dose, year)


ggplot(output2 %>% 
         filter(pfpr==0.1) %>%
         filter(intervention %in% c('EPI', 'seasonal 5-dose', 'hybrid 5-dose')), 
       aes(x=year, y=coverage, color=dose, group=interaction(dose, seasonality))) + 
  geom_line() + 
  facet_grid(vars(seasonality), vars(intervention),  labeller = labeller(seasonality = seas.labs)) + 
  labs(x='Year', y='Coverage', color='Dose') +
  scale_x_continuous(breaks = seq(1,15,1)) + 
  theme_classic() 
  
```

<br>

# 6. Cost-effectiveness: ITNs vs. RTS,S

A continuation of [Pete's paper](https://gh.bmj.com/content/2/1/e000090)? 

```{r cost effectiveness, message=FALSE, warning=FALSE, fig.width=7, fig.height=10, fig.align="center"}
# read in data from HPC runs
output <- readRDS('C:/Users/htopazia/OneDrive - Imperial College London/Github/rtss_malariasimulation/rds/rtss_hybrid_all.rds') 

```

<br>

# Other ideas

* equity in coverage gaps

* malariatools (Tom & Ellie)

* incorporating seasonal births

* catch-up vaccination and flexible start dates

* correlation between seasonality and EPI coverage gaps: [MVIP results](https://cdn.who.int/media/docs/default-source/immunization/mvip/full-evidence-report-on-the-rtss-as01-malaria-vaccine-for-sage-mpag-(sept2021).pdf?sfvrsn=c9737be_5) (pg 49) and [WHO immunisation dashboard](https://immunizationdata.who.int/)

* calculate vaccine efficacy (Risk among unvaccinated group − risk among vaccinated group)  / Risk among unvaccinated group. make outputs for vaccinated vs. unvaccinated cohorts

* look at life-years averted under each strategy. And life-years averted per dose
